Move `StatusBadge` to dedicated module
